+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add `sdwan_cisco_trustsec_feature_template` resource and data source
 - Add `sdwan_cisco_vpn_interface_gre_feature_template` resource and data source
 - Add `sdwan_cellular_profile_feature_template` resource and data source
+- Add `sdwan_cedge_igmp_feature_template` resource and data source
 
 ## 0.3.6